+++ bika/lims/content/analysisrequest.py
@@ -175,7 +175,6 @@ schema = BikaSchema.copy() + Schema((
         'CCContact',
         multiValued=1,
         allowed_types=('Contact',),
-        relationship='AnalysisRequestCCContact',
         mode="rw",
         read_permission=View,
         write_permission=FieldEditContact,
@@ -291,7 +290,6 @@ schema = BikaSchema.copy() + Schema((
     UIDReferenceField(
         'Batch',
         allowed_types=('Batch',),
-        relationship='AnalysisRequestBatch',
         mode="rw",
         read_permission=View,
         write_permission=FieldEditBatch,
@@ -329,7 +327,6 @@ schema = BikaSchema.copy() + Schema((
         'SubGroup',
         required=False,
         allowed_types=('SubGroup',),
-        relationship='AnalysisRequestSubGroup',
         mode="rw",
         read_permission=View,
         write_permission=FieldEditBatch,
@@ -360,7 +357,6 @@ schema = BikaSchema.copy() + Schema((
     UIDReferenceField(
         'Template',
         allowed_types=('ARTemplate',),
-        relationship='AnalysisRequestARTemplate',
         mode="rw",
         read_permission=View,
         write_permission=FieldEditTemplate,
@@ -388,7 +384,6 @@ schema = BikaSchema.copy() + Schema((
         'Profiles',
         multiValued=1,
         allowed_types=('AnalysisProfile',),
-        relationship='AnalysisRequestAnalysisProfiles',
         mode="rw",
         read_permission=View,
         write_permission=FieldEditProfiles,
@@ -618,7 +613,6 @@ schema = BikaSchema.copy() + Schema((
         required=0,
         primary_bound=True,  # field changes propagate to partitions
         allowed_types='AnalysisSpec',
-        relationship='AnalysisRequestAnalysisSpec',
         mode="rw",
         read_permission=View,
         write_permission=FieldEditSpecification,
@@ -669,7 +663,6 @@ schema = BikaSchema.copy() + Schema((
         'PublicationSpecification',
         required=0,
         allowed_types='AnalysisSpec',
-        relationship='AnalysisRequestPublicationSpec',
         mode="rw",
         read_permission=View,
         write_permission=FieldEditPublicationSpecifications,
@@ -977,7 +970,6 @@ schema = BikaSchema.copy() + Schema((
     UIDReferenceField(
         'Invoice',
         allowed_types=('Invoice',),
-        relationship='AnalysisRequestInvoice',
         mode="rw",
         read_permission=View,
         write_permission=ModifyPortalContent,
@@ -1188,7 +1180,6 @@ schema = BikaSchema.copy() + Schema((
     UIDReferenceField(
         "DetachedFrom",
         allowed_types=("AnalysisRequest",),
-        relationship="AnalysisRequestDetachedFrom",
         mode="rw",
         read_permission=View,
         write_permission=ModifyPortalContent,